🔨 Refactor: 새로운 경로 설정
