+++ src/app/lobby/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import useMainSocketConnect from "@/hooks/useMainSocketConnect";
+import useMainSocketConnect from "@/hooks/socket/useMainSocketConnect";
 
 import LobbyCanvas from "./_component/Canvas/Canvas";
